Move ancestors_of into utils
